+++ ChangeLog.md
@@ -28,6 +28,7 @@
 * Prise en compte correcte des malus d'armure et de casques pour les compétences sur la fiche.
 
 ### Autres améliorations
+* Ajout d'une option --saveDM pour éviter tous les dégâts d'une attaque.
 * Prise en compte des options de sélections pour !cof-init.
 * Option d'attaque --forceMinimum
 * Prise en compte de --message pour !cof-animer-mort.